Theodore/proofnarrowing (#34)
* Added proof narrowing feature

Given a proof for a range of leaves [n..m] (and no other information about the tree), allows supplying sub-ranges contiguous with each end of the original range - i.e. [n..k], [l..m], where k < l - to generate a new proof for the sub-range [k..l].

This is particularly useful in Celestia, for generating proofs of specific share ranges within a namespace (e.g. individual proofs for each blob, when there are multiple blobs in a namespace) using a namespace range proof.